requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via discord-py
aiosignal==1.4.0
    # via aiohttp
anyio==4.11.0
    # via httpx
apnggif==0.1.4
    # via -r requirements.in
attrs==25.4.0
    # via
    #   -r requirements.in
    #   aiohttp
beautifulsoup4==4.14.2
    # via -r requirements.in
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
click==8.3.0
    # via apnggif
coloredlogs==15.0.1
    # via apnggif
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
discord-py==2.6.4
    # via -r requirements.in
fonttools==4.60.1
    # via matplotlib
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   httpx
    #   yarl
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.7
    # via -r requirements.in
msgspec==0.19.0
    # via -r requirements.in
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
packaging==25.0
    # via matplotlib
pillow==12.0.0
    # via
    #   -r requirements.in
    #   matplotlib
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
python-dotenv==1.1.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   beautifulsoup4
yarl==1.22.0
    # via aiohttp
